+++ src/early_multicellular_stage/editor/action_data/StartCellTypeEditActionData.cs
@@ -12,7 +12,7 @@ public class StartCellTypeEditActionData : EditorCombinableActionData
         StartedCellTypeEdit = startedCellTypeEdit;
     }
 
-    protected override int CalculateCostInternal()
+    protected override float CalculateCostInternal()
     {
         return 0;
     }
