+++ src/general/IEditorWithActions.cs
@@ -1,6 +1,6 @@
 ﻿public interface IEditorWithActions : IEditor
 {
-    public new int MutationPoints { get; set; }
+    public new float MutationPoints { get; set; }
 
     public void Undo();
     public void Redo();
