+++ src/late_multicellular_stage/editor/MetaballBodyEditorComponent.cs
@@ -372,7 +372,7 @@ public partial class MetaballBodyEditorComponent :
     {
     }
 
-    protected override int CalculateCurrentActionCost()
+    protected override float CalculateCurrentActionCost()
     {
         if (activeActionName == null || !Editor.ShowHover)
             return 0;
